requirements.txt
requests
scikit-learn
sentence-transformers
numpy
streamlit
python-dotenv

